Fixed GraphicEngine distrinction between elements

+++ src/Globals.h
@@ -6,8 +6,8 @@ extern GraphicEngine *graphicEngine;
 
 namespace {
 	
-	extern const int SCREEN_WIDTH = 800;
-	extern const int SCREEN_HEIGHT = 600;
+	extern const float SCREEN_WIDTH = 800;
+	extern const float SCREEN_HEIGHT = 600;
 
 	namespace MenuDesign
 	{
